telegram_pay_bot/requirements.txt
python-telegram-bot==20.7
python-dotenv
Pillow
pytesseract
requests
pymysql
